Add names to Commands
When jsverify fails it prints out the commands that it ran. However, the
commands all looked like empty objects because when it prints to the console it
does not follow the prototype chain. As a hack I simply added a name to each
command so that when the generative test fails I can see what commands ran.